Reorganización de archivos, db y routes listas
