+++ src/config/config.js
@@ -1,25 +1,19 @@
-/* eslint-disable no-undef */
-require('dotenv').config(); // this is important!
+require('dotenv').config();
+
 module.exports = {
-"development": {
-    "username": process.env.DB_USERNAME,
-    "password": process.env.DB_PASSWORD,
-    "database": process.env.DB_DATABASE,
-    "host": process.env.DB_HOST,
-    "dialect": "postgres"
-},
-"test": {
-    "username": "root",
-    "password": null,
-    "database": "database_test",
-    "host": "127.0.0.1",
-    "dialect": "postgres"
-},
-"production": {
-    "username": "root",
-    "password": null,
-    "database": "database_production",
-    "host": "127.0.0.1",
-    "dialect": "postgres"
-}
+  development: {
+    use_env_variable: "PSQL_DATABASE_URL_DEVELOPMENT",
+    dialect: "postgres",
+  },
+  test: {
+    username: "root",
+    password: null,
+    database: "database_test",
+    host: "127.0.0.1",
+    dialect: "postgres",
+  },
+  production: {
+    use_env_variable: "PSQL_DATABASE_URL_PRODUCTION",
+    dialect: "postgres",
+  },
 };
